Split collectives to separate files, re-consider API
