+++ src/gpuarray/array.h
@@ -424,7 +424,7 @@ GPUARRAY_PUBLIC int GpuArray_transpose_inplace(GpuArray *a,
                                               const unsigned int *new_axes);
 
 /**
- * Relase all device and host memory associated with `a`.
+ * Release all device and host memory associated with `a`.
  *
  * This function frees all host memory, and releases the device memory
  * if it is the owner. In case an array has views it is the
@@ -458,7 +458,7 @@ GPUARRAY_PUBLIC int GpuArray_share(const GpuArray *a, const GpuArray *b);
 GPUARRAY_PUBLIC gpucontext *GpuArray_context(const GpuArray *a);
 
 /**
- * Copies all the elements of and array to another.
+ * Copies all the elements of one array to another.
  *
  * The arrays `src` and `dst` must have the same size (total number of
  * elements) and be in the same context.
@@ -606,78 +606,6 @@ GPUARRAY_PUBLIC const char *GpuArray_error(const GpuArray *a, int err);
 GPUARRAY_PUBLIC void GpuArray_fprintf(FILE *fd, const GpuArray *a);
 
 GPUARRAY_PUBLIC int GpuArray_fdump(FILE *fd, const GpuArray *a);
-
-/************************************************************************************
-*                         multi-gpu collectives interface                          *
-************************************************************************************/
-
-/**
- * \brief TODO
- * \param src [const GpuArray*] TODO
- * \param dest [GpuArray*] TODO
- * \param count [int] TODO
- * \param typecode [int] TODO
- * \param opcode [int] TODO
- * \param root [int] TODO
- * \param comm [gpucomm*] TODO
- * \return int TODO
- */
-GPUARRAY_PUBLIC int GpuArray_reduce(const GpuArray* src, GpuArray* dest,
-                                    int count, int typecode, int opcode,
-                                    int root, gpucomm* comm);
-
-/**
- * \brief TODO
- * \param src [const GpuArray*] TODO
- * \param dest [GpuArray*] TODO
- * \param count [int] TODO
- * \param typecode [int] TODO
- * \param opcode [int] TODO
- * \param comm [gpucomm*] TODO
- * \return int TODO
- */
-GPUARRAY_PUBLIC int GpuArray_all_reduce(const GpuArray* src, GpuArray* dest,
-                                        int count, int typecode, int opcode,
-                                        gpucomm* comm);
-
-/**
- * \brief TODO
- * \param src [const GpuArray*] TODO
- * \param dest [GpuArray*] TODO
- * \param count [int] TODO
- * \param typecode [int] TODO
- * \param opcode [int] TODO
- * \param comm [gpucomm*] TODO
- * \return int TODO
- */
-GPUARRAY_PUBLIC int GpuArray_reduce_scatter(const GpuArray* src, GpuArray* dest,
-                                            int count, int typecode, int opcode,
-                                            gpucomm* comm);
-
-/**
- * \brief TODO
- * \param array [GpuArray*] TODO
- * \param count [int] TODO
- * \param typecode [int] TODO
- * \param root [int] TODO
- * \param comm [gpucomm*] TODO
- * \return int TODO
- */
-GPUARRAY_PUBLIC int GpuArray_broadcast(GpuArray* array, int count, int typecode,
-                                       int root, gpucomm* comm);
-
-/**
- * \brief TODO
- * \param src [const GpuArray*] TODO
- * \param dest [GpuArray*] TODO
- * \param count [int] TODO
- * \param typecode [int] TODO
- * \param comm [gpucomm*] TODO
- * \return int TODO
- */
-GPUARRAY_PUBLIC int GpuArray_all_gather(const GpuArray* src, GpuArray* dest,
-                                        int count, int typecode,
-                                        gpucomm* comm);
 
 #ifdef __cplusplus
 }
